# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Balance Alert System API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthStatus {
  authenticated: boolean;
  requiresAuth: boolean;
}

export interface AuthLoginInput {
  otp: string;
}

export interface AuthResult {
  success: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Contact {
  id: number;
  fullName: string;
  phoneNumber: string;
  email: string;
  position: string;
  role: string;
  createdAt: string;
}

export interface ContactInput {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 1 */
  phoneNumber: string;
  email: string;
  position: string;
  role: string;
}

export interface ContactUpdate {
  fullName?: string;
  phoneNumber?: string;
  email?: string;
  position?: string;
  role?: string;
}

export interface Settings {
  id: number;
  smsEnabled: boolean;
  /** @nullable */
  smsApiUrl?: string | null;
  /** @nullable */
  smsBodyTemplate?: string | null;
  smtpEnabled: boolean;
  /** @nullable */
  smtpHost?: string | null;
  /** @nullable */
  smtpPort?: number | null;
  /** @nullable */
  smtpUser?: string | null;
  /** @nullable */
  smtpFrom?: string | null;
  telegramEnabled: boolean;
  /** @nullable */
  telegramBotToken?: string | null;
  /** @nullable */
  telegramChatId?: string | null;
  grafanaUrl: string;
  /** @nullable */
  grafanaApiKey?: string | null;
  /** @nullable */
  grafanaDashboardUid?: string | null;
  scheduleEnabled: boolean;
  scheduleHour: number;
  /** @nullable */
  scheduleCron?: string | null;
  thresholdStaff: number;
  thresholdManager: number;
  thresholdMd: number;
  totpEnabled: boolean;
  /** @nullable */
  totpSecret?: string | null;
  /** @nullable */
  googleSheetUrl?: string | null;
}

export interface SettingsUpdate {
  smsEnabled?: boolean;
  smsApiUrl?: string;
  smsBodyTemplate?: string;
  smtpEnabled?: boolean;
  smtpHost?: string;
  smtpPort?: number;
  smtpUser?: string;
  smtpPassword?: string;
  smtpFrom?: string;
  telegramEnabled?: boolean;
  telegramBotToken?: string;
  telegramChatId?: string;
  grafanaUrl?: string;
  grafanaApiKey?: string;
  grafanaDashboardUid?: string;
  scheduleEnabled?: boolean;
  scheduleHour?: number;
  scheduleCron?: string;
  thresholdStaff?: number;
  thresholdManager?: number;
  thresholdMd?: number;
  totpEnabled?: boolean;
  totpSecret?: string;
  googleSheetUrl?: string;
}

export interface OtpInput {
  otp: string;
}

export interface OtpResult {
  success: boolean;
  token: string;
}

export interface TotpSetup {
  secret: string;
  qrCodeUrl: string;
  otpAuthUrl: string;
}

export interface ClientBalance {
  metric: string;
  remainingBalance: number;
  dailyConsumption: number;
  daysRemaining: number;
  /** Average daily consumption over the last 7 days */
  recentDailyConsumption: number;
  /** Estimated days remaining based on the last 7-day consumption rate (-1 if no data) */
  daysRemainingRecent: number;
  /** Actual SMS consumption yesterday */
  yesterdayConsumption: number;
  /**
   * Percentage change in consumption: yesterday vs day before (null if insufficient data)
   * @nullable
   */
  dailyChangePercent?: number | null;
  severity: string;
  /** @nullable */
  lastUpdated?: string | null;
}

export interface ConsumptionDataPoint {
  day: number;
  date: string;
  consumption: number;
}

export interface ConsumptionHistory {
  metric: string;
  currentMonth: ConsumptionDataPoint[];
  previousMonth: ConsumptionDataPoint[];
  currentMonthLabel: string;
  previousMonthLabel: string;
}

export interface AlertDetail {
  metric: string;
  daysRemaining: number;
  severity: string;
  contactsNotified: number;
}

export interface AlertTriggerResult {
  success: boolean;
  notificationsSent: number;
  errors: string[];
  clientsChecked: number;
  details?: AlertDetail[];
}

export interface AlertHistoryEntry {
  id: number;
  metric: string;
  daysRemaining: number;
  severity: string;
  channel: string;
  recipientCount: number;
  sentAt: string;
  success: boolean;
  /** @nullable */
  errorMessage?: string | null;
}

export interface AlertSummary {
  totalAlertsSent: number;
  clientsInWarning: number;
  clientsInCritical: number;
  clientsInEmergency: number;
  /** @nullable */
  lastRunAt: string | null;
}

export type GetConsumptionHistoryParams = {
  /**
   * Client/org name
   */
  metric: string;
};