# lib/api-zod/src/generated/types/alertHistoryEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Balance Alert System API
 * OpenAPI spec version: 0.1.0
 */

export interface AlertHistoryEntry {
  id: number;
  metric: string;
  daysRemaining: number;
  severity: string;
  channel: string;
  recipientCount: number;
  sentAt: string;
  success: boolean;
  /** @nullable */
  errorMessage?: string | null;
}