# lib/api-zod/src/generated/types/alertTriggerResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Balance Alert System API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertDetail } from "./alertDetail";

export interface AlertTriggerResult {
  success: boolean;
  notificationsSent: number;
  errors: string[];
  clientsChecked: number;
  details?: AlertDetail[];
}