# lib/api-zod/src/generated/types/clientBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Balance Alert System API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientBalance {
  metric: string;
  remainingBalance: number;
  dailyConsumption: number;
  daysRemaining: number;
  /** Average daily consumption over the last 7 days */
  recentDailyConsumption: number;
  /** Estimated days remaining based on the last 7-day consumption rate (-1 if no data) */
  daysRemainingRecent: number;
  /** Actual SMS consumption yesterday */
  yesterdayConsumption: number;
  /**
   * Percentage change in consumption: yesterday vs day before (null if insufficient data)
   * @nullable
   */
  dailyChangePercent?: number | null;
  severity: string;
  /** @nullable */
  lastUpdated?: string | null;
}