# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Balance Alert System API
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  fullName: string;
  phoneNumber: string;
  email: string;
  position: string;
  role: string;
  createdAt: string;
}