# lib/api-zod/src/generated/types/otpResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Balance Alert System API
 * OpenAPI spec version: 0.1.0
 */

export interface OtpResult {
  success: boolean;
  token: string;
}