# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Balance Alert System API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  smsEnabled: boolean;
  /** @nullable */
  smsApiUrl?: string | null;
  /** @nullable */
  smsBodyTemplate?: string | null;
  smtpEnabled: boolean;
  /** @nullable */
  smtpHost?: string | null;
  /** @nullable */
  smtpPort?: number | null;
  /** @nullable */
  smtpUser?: string | null;
  /** @nullable */
  smtpFrom?: string | null;
  telegramEnabled: boolean;
  /** @nullable */
  telegramBotToken?: string | null;
  /** @nullable */
  telegramChatId?: string | null;
  grafanaUrl: string;
  /** @nullable */
  grafanaApiKey?: string | null;
  /** @nullable */
  grafanaDashboardUid?: string | null;
  scheduleEnabled: boolean;
  scheduleHour: number;
  /** @nullable */
  scheduleCron?: string | null;
  thresholdStaff: number;
  thresholdManager: number;
  thresholdMd: number;
  totpEnabled: boolean;
  /** @nullable */
  totpSecret?: string | null;
  /** @nullable */
  googleSheetUrl?: string | null;
}