# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Balance Alert System API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsUpdate {
  smsEnabled?: boolean;
  smsApiUrl?: string;
  smsBodyTemplate?: string;
  smtpEnabled?: boolean;
  smtpHost?: string;
  smtpPort?: number;
  smtpUser?: string;
  smtpPassword?: string;
  smtpFrom?: string;
  telegramEnabled?: boolean;
  telegramBotToken?: string;
  telegramChatId?: string;
  grafanaUrl?: string;
  grafanaApiKey?: string;
  grafanaDashboardUid?: string;
  scheduleEnabled?: boolean;
  scheduleHour?: number;
  scheduleCron?: string;
  thresholdStaff?: number;
  thresholdManager?: number;
  thresholdMd?: number;
  totpEnabled?: boolean;
  totpSecret?: string;
  googleSheetUrl?: string;
}